bot/pyproject.toml
[tool.poetry]
name = "dnd-helper-bot"
version = "0.1.0"
description = "Telegram bot for DnD Helper"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
  { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-telegram-bot = ">=21"
redis = ">=5"
httpx = ">=0.27"

[tool.poetry.group.dev.dependencies]
pytest = ">=8"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"


